major update
gate patterns:

you can now use a pattern as the REPEATS or LENGTH parameter for
patterns.
i.e.:

(pseq '(1 2 3) (pfunc (lambda () *foo*)))

this will continue to output as long as *foo* is > 0. if *foo* is set
to 0, then the pseq pstream will return nils at the end of its list.

this makes things more interactive and "live".

use `pattern-as-pstream` if you want to convert only a pattern to a
pstream, because non-pattern objects converted to pstreams will only
yield one value.