alsa-midi: implement instrument map + various other improvements
to play an event on the midi backend, its :instrument must either be
a number (which will be used as the program number), or a symbol or
string that exists in the *alsa-midi-instrument-map*. the instrument
map can be accessed and altered with the
alsa-midi-instrument-program-number function, which is setf-able. any
events whose :instrument doesn't match that criteria and whose
:backend doesn't contain :alsa-midi will be skipped by the backend.

additionally, get-alsa-midi-cc-mapping and set-alsa-midi-cc-mapping
have been deprecated in favor of alsa-midi-cc-mapping, which is also setf-able.